% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigworld-package.R
\docType{package}
\name{pigworld-package}
\alias{pigworld}
\alias{pigworld-package}
\title{pigworld: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
